requirements.txt
aiohappyeyeballs                               
aiohttp                                      
aiosignal                                      
annotated-types                                
anyio                                   
asgiref                                        
async-timeout                                  
attrs                                       
backoff                                        
bcrypt                                         
build                                   
cachetools                               
certifi                                  
charset-normalizer                       
chroma-hnswlib                           
chromadb                                
click                                    
colorama                                 
coloredlogs                              
dataclasses-json                         
Deprecated                               
distro                                   
durationpy                               
exceptiongroup                           
fastapi                                  
filelock                                 
flatbuffers                              
frozenlist                               
fsspec                                   
google-auth                              
googleapis-common-protos                 
greenlet                                 
grpcio                                   
h11                                      
httpcore                                 
httptools                                
httpx                                    
httpx-sse                                
huggingface-hub                          
humanfriendly                            
idna                                     
importlib_metadata                       
importlib_resources                      
jiter                                    
jsonpatch                                
jsonpointer                              
kubernetes                               
langchain                                
langchain-chroma                         
langchain-community                      
langchain-core                           
langchain-openai                         
langchain-text-splitters                 
langgraph                                
langgraph-checkpoint                     
langgraph-sdk                            
langsmith                                
lark                                     
markdown-it-py                           
marshmallow                              
mdurl                                    
mmh3                                     
monotonic                                
mpmath                                   
msgpack                                  
multidict                                
mypy-extensions                          
numpy                                    
oauthlib                                 
onnxruntime                              
openai                                   
opentelemetry-api                        
opentelemetry-exporter-otlp-proto-common 
opentelemetry-exporter-otlp-proto-grpc   
opentelemetry-instrumentation            
opentelemetry-instrumentation-asgi       
opentelemetry-instrumentation-fastapi    
opentelemetry-proto                      
opentelemetry-sdk                        
opentelemetry-semantic-conventions      
opentelemetry-util-http                  
orjson                                   
overrides                                
packaging                                
pandas                                   
pip                                      
posthog                                  
propcache                                
protobuf                                 
pyasn1                                   
pyasn1_modules                           
pydantic                                 
pydantic_core                            
pydantic-settings                        
Pygments                                 
PyPika                                   
pyproject_hooks                          
pyreadline3                              
python-dateutil                          
python-dotenv                            
pytz                                     
PyYAML                                   
regex                                    
requests                                 
requests-oauthlib                        
requests-toolbelt                        
rich                                     
rsa                                      
setuptools                               
shellingham                              
six                                      
sniffio                                  
SQLAlchemy                               
starlette                                
sympy                                    
tavily-python                            
tenacity                                 
tiktoken                                 
tokenizers                               
tomli                                    
tqdm                                     
typer                                    
typing_extensions                        
typing-inspect                           
tzdata                                   
urllib3                                  
uvicorn                                  
watchfiles                               
websocket-client                         
websockets                               
wrapt                                    
yarl                                     
zipp                                     